requirements.txt
yaml
torch
torchvision
os
pillow
numpy
visdom
mlflow